+++ yaylib/client.py
@@ -30,7 +30,6 @@ import random
 import logging
 
 from datetime import datetime
-from enum import Enum
 from typing import Optional, List, Dict, Any
 
 import httpx
@@ -124,6 +123,7 @@ from .responses import (
     UsersByTimestampResponse,
     UserTimestampResponse,
 )
+from .types import PolicyType
 from .utils import Colors, generate_jwt
 from .ws import Intents, WebSocketInteractor
 
@@ -134,75 +134,6 @@ except ImportError:
 
 
 __all__ = "Client"
-
-
-class PostType(Enum):
-    """投稿の種類"""
-
-    text = "text"
-    """ `text`: テキストのみ投稿タイプ"""
-    media = "media"
-    """ `media`: メディアを含める投稿タイプ"""
-    image = "image"
-    """ `image`: 画像を含める投稿タイプ"""
-    video = "video"
-    """ `video`: ビデオを含める投稿タイプ"""
-    survey = "survey"
-    """ `survey`: アンケートを含める投稿タイプ"""
-    call = "call"
-    """ `call`: 通話用の投稿タイプ"""
-    shareable_url = "shareable_url"
-    """ `shareable_url`: サークルやスレッド共有用の投稿タイプ"""
-
-
-class CallType(Enum):
-    """通話の種類"""
-
-    voice = "voice"
-    """ `voice`: 音声通話用の通話タイプ"""
-    video = "vdo"
-    """ `vdo`: ビデオ通話用の通話タイプ"""
-
-
-class ImageType(Enum):
-    """画像の種類"""
-
-    post = "post"
-    """ `post`: 投稿に画像をアップロードする際の画像タイプ"""
-    chat_message = "chat_message"
-    """ `chat_message`: 個人チャットに画像をアップロードする際の画像タイプ"""
-    chat_background = "chat_background"
-    """ `chat_background`: 個人チャットの背景用に画像をアップロードする際の画像タイプ"""
-    report = "report"
-    """ `report`: 通報用の画像をアップロードする際の画像タイプ"""
-    user_avatar = "user_avatar"
-    """ `user_avatar`: プロフィール画像をアップロードする際の画像タイプ"""
-    user_cover = "user_cover"
-    """ `user_cover`: プロフィールの背景画像をアップロードする際の画像タイプ"""
-    group_cover = "group_cover"
-    """ `group_cover`: グループの背景画像をアップロードする際の画像タイプ"""
-    group_thread_icon = "group_thread_icon"
-    """ `group_thread_icon`: グループ内のスレッド用アイコンをアップロードする際の画像タイプ"""
-    group_icon = "group_icon"
-    """ `group_icon`: グループのアイコンをアップロードする際の画像タイプ"""
-
-
-class ShareableType(Enum):
-    """共有の種類"""
-
-    group = "group"
-    """ `group`: サークル用の共有タイプ"""
-    thread = "thread"
-    """ `thread`: スレッド用の共有タイプ"""
-
-
-class PolicyType(Enum):
-    """利用規約の種類"""
-
-    privacy_policy = "privacy_policy"
-    """ `privacy_policy`: プライバシーポリシー"""
-    terms_of_use = "terms_of_use"
-    """ `privacy_policy`: 利用規約"""
 
 
 class HeaderInterceptor(object):
